Added conditional logic to Podfile.
Altered Podfile to use local BlueRocketFuelCore project if 'pod install' is run from BlueRocketFuelApp directory.  In that case it is assumed that the developer is working on the Fuel framework itself.  However, if the current working directory is not called BlueRocketFuelApp, then it is assumed that the developer used launchNewProject.sh to spawn a new project.  In that case, CocoaPods downloads the BlueRocketFuelCore project from Github instead.